+++ BlueprintUI/Sources/Element/MeasureElementStorage.swift
@@ -31,7 +31,7 @@ extension MeasureElementStorage {
         )
     }
 
-    func performCaffeinatedLayout(
+    func performLayout(
         frame: CGRect,
         environment: Environment,
         node: LayoutTreeNode